# gen/http/books/server/encode_decode.go
// Code generated by goa v3.25.3, DO NOT EDIT.
//
// books HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/Velin-Todorov/zetta-task/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"
	"strconv"
	"unicode/utf8"

	books "github.com/Velin-Todorov/zetta-task/gen/books"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeGetBooksResponse returns an encoder for responses returned by the
// books getBooks endpoint.
func EncodeGetBooksResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*books.Book)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "application/json")
		enc := encoder(ctx, w)
		body := NewGetBooksResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetBooksRequest returns a decoder for requests sent to the books
// getBooks endpoint.
func DecodeGetBooksRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*books.GetBooksPayload, error) {
	return func(r *http.Request) (*books.GetBooksPayload, error) {
		var payload *books.GetBooksPayload
		var (
			title           *string
			author          *string
			publishedAt     *string
			publishedAfter  *string
			publishedBefore *string
			limit           *int64
			offset          *int64
			err             error
		)
		qp := r.URL.Query()
		titleRaw := qp.Get("title")
		if titleRaw != "" {
			title = &titleRaw
		}
		if title != nil {
			if utf8.RuneCountInString(*title) < 1 {
				err = goa.MergeErrors(err, goa.InvalidLengthError("title", *title, utf8.RuneCountInString(*title), 1, true))
			}
		}
		authorRaw := qp.Get("author")
		if authorRaw != "" {
			author = &authorRaw
		}
		if author != nil {
			if utf8.RuneCountInString(*author) < 1 {
				err = goa.MergeErrors(err, goa.InvalidLengthError("author", *author, utf8.RuneCountInString(*author), 1, true))
			}
		}
		publishedAtRaw := qp.Get("published_at")
		if publishedAtRaw != "" {
			publishedAt = &publishedAtRaw
		}
		if publishedAt != nil {
			err = goa.MergeErrors(err, goa.ValidateFormat("published_at", *publishedAt, goa.FormatDate))
		}
		publishedAfterRaw := qp.Get("published_after")
		if publishedAfterRaw != "" {
			publishedAfter = &publishedAfterRaw
		}
		if publishedAfter != nil {
			err = goa.MergeErrors(err, goa.ValidateFormat("published_after", *publishedAfter, goa.FormatDate))
		}
		publishedBeforeRaw := qp.Get("published_before")
		if publishedBeforeRaw != "" {
			publishedBefore = &publishedBeforeRaw
		}
		if publishedBefore != nil {
			err = goa.MergeErrors(err, goa.ValidateFormat("published_before", *publishedBefore, goa.FormatDate))
		}
		{
			limitRaw := qp.Get("limit")
			if limitRaw != "" {
				v, err2 := strconv.ParseInt(limitRaw, 10, 64)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("limit", limitRaw, "integer"))
				}
				limit = &v
			}
		}
		{
			offsetRaw := qp.Get("offset")
			if offsetRaw != "" {
				v, err2 := strconv.ParseInt(offsetRaw, 10, 64)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("offset", offsetRaw, "integer"))
				}
				offset = &v
			}
		}
		if err != nil {
			return payload, err
		}
		payload = NewGetBooksPayload(title, author, publishedAt, publishedAfter, publishedBefore, limit, offset)

		return payload, nil
	}
}

// EncodeGetBookResponse returns an encoder for responses returned by the books
// getBook endpoint.
func EncodeGetBookResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*books.Book)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "application/json")
		enc := encoder(ctx, w)
		body := NewGetBookResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetBookRequest returns a decoder for requests sent to the books
// getBook endpoint.
func DecodeGetBookRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*books.GetBookPayload, error) {
	return func(r *http.Request) (*books.GetBookPayload, error) {
		var payload *books.GetBookPayload
		var (
			id  int64
			err error

			params = mux.Vars(r)
		)
		{
			idRaw := params["id"]
			v, err2 := strconv.ParseInt(idRaw, 10, 64)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("id", idRaw, "integer"))
			}
			id = v
		}
		if err != nil {
			return payload, err
		}
		payload = NewGetBookPayload(id)

		return payload, nil
	}
}

// EncodeCreateBookResponse returns an encoder for responses returned by the
// books createBook endpoint.
func EncodeCreateBookResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*books.Book)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "application/json")
		enc := encoder(ctx, w)
		body := NewCreateBookResponseBody(res)
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeCreateBookRequest returns a decoder for requests sent to the books
// createBook endpoint.
func DecodeCreateBookRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*books.CreateBookPayload, error) {
	return func(r *http.Request) (*books.CreateBookPayload, error) {
		var payload *books.CreateBookPayload
		var (
			body CreateBookRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return payload, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return payload, gerr
			}
			return payload, goa.DecodePayloadError(err.Error())
		}
		err = ValidateCreateBookRequestBody(&body)
		if err != nil {
			return payload, err
		}
		payload = NewCreateBookPayload(&body)

		return payload, nil
	}
}

// EncodeCreateBookCoverResponse returns an encoder for responses returned by
// the books createBookCover endpoint.
func EncodeCreateBookCoverResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*books.Book)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "application/json")
		enc := encoder(ctx, w)
		body := NewCreateBookCoverResponseBody(res)
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeCreateBookCoverRequest returns a decoder for requests sent to the
// books createBookCover endpoint.
func DecodeCreateBookCoverRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*books.CreateBookCoverPayload, error) {
	return func(r *http.Request) (*books.CreateBookCoverPayload, error) {
		var payload *books.CreateBookCoverPayload
		var (
			id  int64
			err error

			params = mux.Vars(r)
		)
		{
			idRaw := params["id"]
			v, err2 := strconv.ParseInt(idRaw, 10, 64)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("id", idRaw, "integer"))
			}
			id = v
		}
		if err != nil {
			return payload, err
		}
		payload = NewCreateBookCoverPayload(id)

		return payload, nil
	}
}

// EncodeUpdateBookResponse returns an encoder for responses returned by the
// books updateBook endpoint.
func EncodeUpdateBookResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*books.Book)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "application/json")
		enc := encoder(ctx, w)
		body := NewUpdateBookResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeUpdateBookRequest returns a decoder for requests sent to the books
// updateBook endpoint.
func DecodeUpdateBookRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*books.UpdateBookPayload, error) {
	return func(r *http.Request) (*books.UpdateBookPayload, error) {
		var payload *books.UpdateBookPayload
		var (
			body UpdateBookRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return payload, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return payload, gerr
			}
			return payload, goa.DecodePayloadError(err.Error())
		}
		err = ValidateUpdateBookRequestBody(&body)
		if err != nil {
			return payload, err
		}

		var (
			id int64

			params = mux.Vars(r)
		)
		{
			idRaw := params["id"]
			v, err2 := strconv.ParseInt(idRaw, 10, 64)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("id", idRaw, "integer"))
			}
			id = v
		}
		if err != nil {
			return payload, err
		}
		payload = NewUpdateBookPayload(&body, id)

		return payload, nil
	}
}

// EncodeUpdateBookCoverResponse returns an encoder for responses returned by
// the books updateBookCover endpoint.
func EncodeUpdateBookCoverResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*books.Book)
		enc := encoder(ctx, w)
		body := NewUpdateBookCoverResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeUpdateBookCoverRequest returns a decoder for requests sent to the
// books updateBookCover endpoint.
func DecodeUpdateBookCoverRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*books.UpdateBookCoverPayload, error) {
	return func(r *http.Request) (*books.UpdateBookCoverPayload, error) {
		var payload *books.UpdateBookCoverPayload
		var (
			id  int64
			err error

			params = mux.Vars(r)
		)
		{
			idRaw := params["id"]
			v, err2 := strconv.ParseInt(idRaw, 10, 64)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("id", idRaw, "integer"))
			}
			id = v
		}
		if err != nil {
			return payload, err
		}
		payload = NewUpdateBookCoverPayload(id)

		return payload, nil
	}
}

// EncodeDeleteBookResponse returns an encoder for responses returned by the
// books deleteBook endpoint.
func EncodeDeleteBookResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		w.WriteHeader(http.StatusNoContent)
		return nil
	}
}

// DecodeDeleteBookRequest returns a decoder for requests sent to the books
// deleteBook endpoint.
func DecodeDeleteBookRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*books.DeleteBookPayload, error) {
	return func(r *http.Request) (*books.DeleteBookPayload, error) {
		var payload *books.DeleteBookPayload
		var (
			id  int64
			err error

			params = mux.Vars(r)
		)
		{
			idRaw := params["id"]
			v, err2 := strconv.ParseInt(idRaw, 10, 64)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("id", idRaw, "integer"))
			}
			id = v
		}
		if err != nil {
			return payload, err
		}
		payload = NewDeleteBookPayload(id)

		return payload, nil
	}
}

// marshalBooksBookToBookResponse builds a value of type *BookResponse from a
// value of type *books.Book.
func marshalBooksBookToBookResponse(v *books.Book) *BookResponse {
	res := &BookResponse{
		ID:          v.ID,
		Title:       v.Title,
		Author:      v.Author,
		CoverURL:    v.CoverURL,
		PublishedAt: v.PublishedAt,
	}

	return res
}