# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Bookstore API","description":"An api that returns data about books in a bookstore","version":"1.0.0"},"servers":[{"url":"http://localhost:80","description":"Default server for bookstore"}],"paths":{"/books":{"get":{"tags":["books"],"summary":"getBooks books","operationId":"books#getBooks","parameters":[{"name":"title","in":"query","allowEmptyValue":true,"schema":{"type":"string","example":"0","minLength":1},"example":"o57"},{"name":"author","in":"query","allowEmptyValue":true,"schema":{"type":"string","example":"4gj","minLength":1},"example":"orm"},{"name":"published_at","in":"query","allowEmptyValue":true,"schema":{"type":"string","example":"1991-11-24","format":"date"},"example":"2002-06-04"},{"name":"published_after","in":"query","allowEmptyValue":true,"schema":{"type":"string","example":"1984-10-22","format":"date"},"example":"2006-10-27"},{"name":"published_before","in":"query","allowEmptyValue":true,"schema":{"type":"string","example":"2007-06-12","format":"date"},"example":"1975-11-03"},{"name":"limit","in":"query","allowEmptyValue":true,"schema":{"type":"integer","example":1658808341667584976,"format":"int64"},"example":7323195020730466378},{"name":"offset","in":"query","allowEmptyValue":true,"schema":{"type":"integer","example":594581378109610628,"format":"int64"},"example":8962451874647435356}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Book"},"example":[{"author":"Veritatis maxime occaecati magni et.","cover_url":"Et sed voluptate nemo voluptates.","id":5133803869587210357,"published_at":"1979-10-09","title":"Nihil accusamus."},{"author":"Veritatis maxime occaecati magni et.","cover_url":"Et sed voluptate nemo voluptates.","id":5133803869587210357,"published_at":"1979-10-09","title":"Nihil accusamus."},{"author":"Veritatis maxime occaecati magni et.","cover_url":"Et sed voluptate nemo voluptates.","id":5133803869587210357,"published_at":"1979-10-09","title":"Nihil accusamus."}]},"example":[{"author":"Veritatis maxime occaecati magni et.","cover_url":"Et sed voluptate nemo voluptates.","id":5133803869587210357,"published_at":"1979-10-09","title":"Nihil accusamus."},{"author":"Veritatis maxime occaecati magni et.","cover_url":"Et sed voluptate nemo voluptates.","id":5133803869587210357,"published_at":"1979-10-09","title":"Nihil accusamus."},{"author":"Veritatis maxime occaecati magni et.","cover_url":"Et sed voluptate nemo voluptates.","id":5133803869587210357,"published_at":"1979-10-09","title":"Nihil accusamus."}]}}}}},"post":{"tags":["books"],"summary":"createBook books","operationId":"books#createBook","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookRequestBody"},"example":{"author":"Qui eaque sequi.","published_at":"2009-03-26","title":"Aliquam excepturi velit incidunt iste molestiae."}}}},"responses":{"201":{"description":"Created response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"},"example":{"author":"Qui ea rerum ut soluta.","cover_url":"Laboriosam voluptates placeat.","id":5913967038983033047,"published_at":"1998-11-25","title":"Quia amet rerum magni."}}}}}}},"/books/{id}":{"delete":{"tags":["books"],"summary":"deleteBook books","operationId":"books#deleteBook","parameters":[{"name":"id","in":"path","description":"ID of the book","required":true,"schema":{"type":"integer","description":"ID of the book","example":5886920984786250176,"format":"int64"},"example":8373621198686456346}],"responses":{"204":{"description":"No Content response."}}},"get":{"tags":["books"],"summary":"getBook books","operationId":"books#getBook","parameters":[{"name":"id","in":"path","description":"ID of the book","required":true,"schema":{"type":"integer","description":"ID of the book","example":2513408602841306082,"format":"int64"},"example":3608411315989402758}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"},"example":{"author":"Laudantium architecto esse.","cover_url":"Blanditiis provident dolor dolor reiciendis.","id":6632813022759676477,"published_at":"1977-12-15","title":"Et numquam."}}}}}},"patch":{"tags":["books"],"summary":"updateBook books","operationId":"books#updateBook","parameters":[{"name":"id","in":"path","description":"ID of the book","required":true,"schema":{"type":"integer","description":"ID of the book","example":618102801709849111,"format":"int64"},"example":1746000879518224380}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBookRequestBody"},"example":{"author":"Aliquam aperiam.","published_at":"1993-02-06","title":"Ea aut sequi autem."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"},"example":{"author":"Corrupti ad eum rerum non ut.","cover_url":"Id non ullam voluptatum.","id":5718424733436786398,"published_at":"2010-11-21","title":"Fugiat perferendis ea qui eum voluptate occaecati."}}}}}}},"/books/{id}/cover":{"post":{"tags":["books"],"summary":"createBookCover books","operationId":"books#createBookCover","parameters":[{"name":"id","in":"path","description":"ID of the book","required":true,"schema":{"type":"integer","description":"ID of the book","example":2503764079057987132,"format":"int64"},"example":4517506437017472857}],"responses":{"201":{"description":"Created response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"},"example":{"author":"Voluptatem aut qui iure in.","cover_url":"Vitae qui sed vel.","id":9072270161687811363,"published_at":"2012-01-07","title":"Consequatur voluptatem temporibus eum doloremque."}}}}}},"put":{"tags":["books"],"summary":"updateBookCover books","operationId":"books#updateBookCover","parameters":[{"name":"id","in":"path","description":"ID of the book","required":true,"schema":{"type":"integer","description":"ID of the book","example":2842795117314359702,"format":"int64"},"example":1890328457544575904}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"},"example":{"author":"Eligendi labore sapiente voluptate inventore nisi.","cover_url":"Mollitia soluta rem.","id":922662716789829762,"published_at":"1979-01-15","title":"Optio quaerat enim consequatur rem pariatur."}}}}}}}},"components":{"schemas":{"Book":{"type":"object","properties":{"author":{"type":"string","example":"Debitis blanditiis ab in dolores ipsa vitae."},"cover_url":{"type":"string","example":"Aut ipsam veritatis maxime."},"id":{"type":"integer","example":1815354530930689837,"format":"int64"},"published_at":{"type":"string","example":"1997-02-17","format":"date"},"title":{"type":"string","example":"Nihil molestiae tempore voluptas ab."}},"description":"A single book","example":{"author":"Repellat natus odit possimus fugit rerum.","cover_url":"Et officiis incidunt.","id":1822507689538841978,"published_at":"1987-12-02","title":"Tempore nihil quisquam voluptatem omnis aliquid."},"required":["id","title","author","published_at"]},"CreateBookRequestBody":{"type":"object","properties":{"author":{"type":"string","description":"Author of the book","example":"Ipsam voluptate error nisi cupiditate adipisci ipsum."},"published_at":{"type":"string","description":"Publication time of the book","example":"2009-02-11","format":"date"},"title":{"type":"string","description":"Title of the book","example":"Repellendus quae ullam iusto est aut."}},"example":{"author":"Quos repellat.","published_at":"1982-02-19","title":"Ratione at dolorem minima odit sed."},"required":["title","author","published_at"]},"UpdateBookRequestBody":{"type":"object","properties":{"author":{"type":"string","description":"Author of the book","example":"Vel voluptatem molestiae."},"published_at":{"type":"string","description":"Publication time of the book","example":"1978-08-02","format":"date"},"title":{"type":"string","description":"Title of the book","example":"Voluptate consequatur sint voluptatibus."}},"example":{"author":"Rem aut blanditiis maiores.","published_at":"1988-11-26","title":"Commodi tenetur dolorem."}}}},"tags":[{"name":"books","description":"Service that returns data about books"}]}