# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
	"time"
)

type Book struct {
	ID          int64
	Title       string
	Author      string
	CoverPath   sql.NullString
	PublishedAt time.Time
}